backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==7.6.10
defusedxml==0.7.1
diff-match-patch==20241021
distlib==0.3.9
Django==5.1.5
django-cleanup==8.1.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-easy-audit==1.3.7
django-extensions==3.2.3
django-filter==23.5
django-health-check==3.18.3
django-import-export==3.3.9
django-polymorphic==3.1.0
django-rest-passwordreset==1.5.0
django-rest-polymorphic==0.1.10
django-restql==0.15.4
django-richtextfield==1.6.2
django-session-timeout==0.1.0
django-simple-history==3.7.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.12.1
drf-yasg==1.21.8
et_xmlfile==2.0.0
filelock==3.17.0
gevent==23.9.1
greenlet==3.1.1
gunicorn==22.0.0
hiredis==3.1.0
identify==2.6.6
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupPy==1.14
mypy-extensions==1.0.0
nodeenv==1.9.1
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
psutil==5.9.8
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
pydotplus==2.0.2
PyJWT==2.10.1
pyparsing==3.2.1
pyPEG2==2.15.2
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
rpds-py==0.22.3
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.1
whitenoise==6.8.2
xlrd==2.0.1
xlwt==1.3.0
zope.event==5.0
zope.interface==7.2
uvicorn~=0.34.0
pytz~=2024.2
